LayoutTests/inspector/audit/basic-debugger-expected.txt

Testing the basic functionality of audits.


== Running test suite: Audit.Basic
-- Running test case: Audit.Basic.Debugger
Testing `debugger`...
PASS: Should not pause on the debugger statement.
PASS: Result should be "pass".
